# swagger.json
{
 "swagger": "2.0",
 "info": {
  "version": "0.0.1",
  "title": "Timeline"
 },
 "host": "localhost:3000",
 "basePath": "/",
 "schemes": [
  "http"
 ],
 "consumes": [
  "application/json"
 ],
 "produces": [
  "application/json"
 ],
 "paths": {
  "/api/init-load-data": {
   "get": {
    "tags": [
     "Init load data"
    ],
    "description": "init load data",
    "parameters": [
     {
      "in": "query",
      "name": "login",
      "description": "login of user",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful response",
      "schema": {
       "title": "data",
       "type": "object",
       "properties": {
        "employees": {
         "type": "array",
         "items": {
          "type": "object"
         }
        },
        "projects": {
         "type": "array",
         "items": {
          "type": "object"
         }
        }
       }
      }
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/auth-user": {
   "post": {
    "tags": [
     "Authenticate user"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "User",
      "description": "Auth user.",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "login": {
         "type": "string"
        },
        "passwd": {
         "type": "string"
        }
       }
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful response",
      "schema": {
       "title": "token",
       "type": "string"
      }
     },
     "400": {
      "description": "Bad request",
      "schema": {
       "title": "body",
       "type": "string"
      }
     },
     "404": {
      "description": "No such user",
      "schema": {
       "title": "body",
       "type": "string"
      }
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/registration-new-user": {
   "post": {
    "tags": [
     "Add new user"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "User",
      "description": "Add new user.",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "login": {
         "type": "string"
        },
        "passwd": {
         "type": "string"
        }
       }
      }
     }
    ],
    "responses": {
     "200": {
      "description": "You have added a new user",
      "schema": {
       "title": "add user",
       "type": "string"
      }
     },
     "201": {
      "description": "This user have already exists",
      "schema": {
       "title": "exists user",
       "type": "string"
      }
     },
     "400": {
      "description": "Bad request",
      "schema": {
       "title": "body",
       "type": "string"
      }
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/add-new-employee": {
   "post": {
    "tags": [
     "Add new employee"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "Employee",
      "description": "Add employee",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "login": {
         "type": "string"
        },
        "name": {
         "type": "string"
        }
       }
      }
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "You have added a new employee",
      "schema": {
       "type": "array",
       "items": {
        "type": "object",
        "properties": {
         "id": {
          "type": "integer",
          "description": "The employee ID."
         },
         "name": {
          "type": "string",
          "description": "The employee name."
         },
         "data": {
          "type": "array",
          "description": "array with filled days",
          "items": {
           "type": "object"
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request",
      "schema": {
       "title": "body",
       "type": "string"
      }
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/add-new-project": {
   "post": {
    "tags": [
     "Add new project"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "Project",
      "description": "Add project",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "login": {
         "type": "string"
        },
        "name": {
         "type": "string"
        },
        "color": {
         "type": "string"
        }
       }
      }
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "You have added a new project",
      "schema": {
       "type": "array",
       "items": {
        "type": "object",
        "properties": {
         "id": {
          "type": "integer",
          "description": "The project ID."
         },
         "name": {
          "type": "string",
          "description": "The project name."
         },
         "color": {
          "type": "string",
          "description": "color project"
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request",
      "schema": {
       "title": "body",
       "type": "string"
      }
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/add-new-fill-day": {
   "post": {
    "tags": [
     "Add new fill day"
    ],
    "parameters": [
     {
      "in": "body",
      "name": "Project",
      "description": "Add fill day",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "login": {
         "type": "string"
        },
        "id": {
         "type": "string",
         "description": "employee ID"
        },
        "number": {
         "type": "integer"
        },
        "day": {
         "type": "string"
        },
        "month": {
         "type": "object",
         "properties": {
          "name": {
           "type": "string",
           "description": "month name"
          },
          "number": {
           "type": "integer",
           "description": "month number"
          }
         }
        },
        "status": {
         "type": "string"
        },
        "color": {
         "type": "string"
        },
        "hours": {
         "type": "integer"
        },
        "year": {
         "type": "integer"
        }
       }
      }
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "You have added a new fill day",
      "schema": {
       "type": "array",
       "items": {
        "type": "object",
        "properties": {
         "id": {
          "type": "integer",
          "description": "The employee ID."
         },
         "name": {
          "type": "string",
          "description": "The employee name."
         },
         "data": {
          "type": "array",
          "description": "array with filled days",
          "items": {
           "type": "object"
          }
         }
        }
       }
      }
     },
     "304": {
      "description": "Day not modified",
      "schema": {
       "type": "object"
      }
     },
     "400": {
      "description": "Bad request",
      "schema": {
       "title": "body",
       "type": "string"
      }
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/save-name-employee/{id}": {
   "put": {
    "tags": [
     "Save new employee name"
    ],
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "type": "string",
      "description": "The employee ID"
     },
     {
      "in": "body",
      "name": "Employee",
      "description": "Employee name",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "name": {
         "type": "string"
        }
       }
      }
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful",
      "schema": {
       "type": "array",
       "items": {
        "type": "object",
        "properties": {
         "id": {
          "type": "integer",
          "description": "The employee ID"
         },
         "name": {
          "type": "string",
          "description": "The employee name"
         },
         "data": {
          "type": "array",
          "description": "array with filled days",
          "items": {
           "type": "object"
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request"
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/save-name-project/{id}": {
   "put": {
    "tags": [
     "Save new project name"
    ],
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "type": "string",
      "description": "The project ID"
     },
     {
      "in": "body",
      "name": "Project",
      "description": "Project name",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "login": {
         "type": "string"
        },
        "name": {
         "type": "string"
        }
       }
      }
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful",
      "schema": {
       "type": "object",
       "properties": {
        "project": {
         "type": "object",
         "properties": {
          "_id": {
           "type": "integer",
           "description": "The project ID"
          },
          "name": {
           "type": "string",
           "description": "The project name"
          },
          "color": {
           "type": "string",
           "description": "project color"
          }
         }
        },
        "employee": {
         "type": "object",
         "properties": {
          "id": {
           "type": "integer",
           "description": "The employee ID"
          },
          "name": {
           "type": "string",
           "description": "The employee name"
          },
          "data": {
           "type": "array",
           "description": "array with filled days",
           "items": {
            "type": "object"
           }
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request"
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/save-hours/{id}": {
   "put": {
    "tags": [
     "Save hours"
    ],
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "type": "string",
      "description": "The day ID"
     },
     {
      "in": "body",
      "name": "Day",
      "description": "Hours in the day",
      "required": true,
      "schema": {
       "type": "object",
       "properties": {
        "login": {
         "type": "string"
        },
        "id": {
         "type": "string",
         "description": "Employee ID"
        },
        "value": {
         "type": "string",
         "description": "New hours value"
        }
       }
      }
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful",
      "schema": {
       "type": "object",
       "properties": {
        "_id": {
         "type": "integer",
         "description": "The employee ID"
        },
        "name": {
         "type": "string",
         "description": "The employee name"
        },
        "data": {
         "type": "array",
         "description": "array with filled days",
         "items": {
          "type": "object"
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request"
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/remove-fill-day/{id}": {
   "delete": {
    "tags": [
     "Remove new fill day"
    ],
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "type": "string",
      "description": "The day ID"
     },
     {
      "in": "query",
      "name": "login",
      "description": "login of user",
      "required": true,
      "type": "string"
     },
     {
      "in": "query",
      "name": "employee",
      "description": "employee ID",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful",
      "schema": {
       "type": "array",
       "items": {
        "type": "object",
        "properties": {
         "id": {
          "type": "integer",
          "description": "The employee ID."
         },
         "name": {
          "type": "string",
          "description": "The employee name."
         },
         "data": {
          "type": "array",
          "description": "array with filled days",
          "items": {
           "type": "object"
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request"
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/remove-employee/{id}": {
   "delete": {
    "tags": [
     "Remove employee"
    ],
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "type": "string",
      "description": "The employee ID"
     },
     {
      "in": "query",
      "name": "login",
      "description": "login of user",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful",
      "schema": {
       "type": "array",
       "items": {
        "type": "object",
        "properties": {
         "_id": {
          "type": "integer",
          "description": "The employee ID."
         },
         "name": {
          "type": "string",
          "description": "The employee name."
         },
         "data": {
          "type": "array",
          "description": "array with filled days",
          "items": {
           "type": "object"
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request"
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  },
  "/api/remove-project/{id}": {
   "delete": {
    "tags": [
     "Remove project"
    ],
    "parameters": [
     {
      "in": "path",
      "name": "id",
      "required": true,
      "type": "string",
      "description": "The project ID"
     },
     {
      "in": "query",
      "name": "login",
      "description": "login of user",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "x-access-token",
      "description": "jwt token",
      "required": true,
      "type": "string"
     },
     {
      "in": "header",
      "name": "Content-Type",
      "description": "Content Type",
      "required": true,
      "type": "string"
     }
    ],
    "responses": {
     "200": {
      "description": "Successful",
      "schema": {
       "type": "object",
       "properties": {
        "projects": {
         "type": "array",
         "items": {
          "type": "object",
          "properties": {
           "_id": {
            "type": "integer",
            "description": "The project ID"
           },
           "name": {
            "type": "string",
            "description": "The project name"
           },
           "color": {
            "type": "string",
            "description": "project color"
           }
          }
         }
        },
        "employees": {
         "type": "array",
         "items": {
          "type": "object",
          "properties": {
           "_id": {
            "type": "integer",
            "description": "The employee ID."
           },
           "name": {
            "type": "string",
            "description": "The employee name."
           },
           "data": {
            "type": "array",
            "description": "array with filled days",
            "items": {
             "type": "object"
            }
           }
          }
         }
        }
       }
      }
     },
     "400": {
      "description": "Bad request"
     },
     "401": {
      "description": "Unauthorized"
     },
     "405": {
      "description": "Method Not Allowed"
     },
     "500": {
      "description": "Internal Server Error"
     }
    }
   }
  }
 }
}